requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
alembic
sqladmin